+++ mayan/apps/lock_manager/backends/model_lock.py
@@ -5,8 +5,7 @@ from .base import LockingBackend
 
 class ModelLock(LockingBackend):
     @classmethod
-    def acquire_lock(cls, name, timeout=None):
-        super(ModelLock, cls).acquire_lock(name=name, timeout=timeout)
+    def _acquire_lock(cls, name, timeout):
         Lock = apps.get_model(app_label='lock_manager', model_name='Lock')
         return ModelLock(
             model_instance=Lock.objects.acquire_lock(
@@ -15,8 +14,7 @@ class ModelLock(LockingBackend):
         )
 
     @classmethod
-    def purge_locks(cls):
-        super(ModelLock, cls).purge_locks()
+    def _purge_locks(cls):
         Lock = apps.get_model(app_label='lock_manager', model_name='Lock')
         Lock.objects.select_for_update().delete()
 
@@ -24,6 +22,5 @@ class ModelLock(LockingBackend):
         self.model_instance = model_instance
         self.name = model_instance.name
 
-    def release(self):
-        super(ModelLock, self).release()
+    def _release(self):
         self.model_instance.release()