+++ mayan/apps/lock_manager/literals.py
@@ -1,3 +1,7 @@
 DEFAULT_LOCK_MANAGER_BACKEND = 'mayan.apps.lock_manager.backends.file_lock.FileLock'
 DEFAULT_LOCK_MANAGER_BACKEND_ARGUMENTS = {}
 DEFAULT_LOCK_MANAGER_DEFAULT_LOCK_TIMEOUT = 30
+
+PURGE_LOCKS_COMMAND = 'purgelocks'
+
+TEST_LOCK_NAME = '_mayan_test_lock'