+++ mayan/apps/lock_manager/tests/test_backends.py
@@ -7,15 +7,14 @@ from django.utils.module_loading import import_string
 from mayan.apps.testing.tests.base import BaseTestCase
 
 from ..exceptions import LockError
+from ..settings import setting_default_lock_timeout
 
-TEST_LOCK_1 = 'test lock 1'
+from .literals import TEST_LOCK_1, TEST_LOCK_LONG_NAME
 
 
-class FileLockTestCase(BaseTestCase):
-    backend_string = 'mayan.apps.lock_manager.backends.file_lock.FileLock'
-
+class BaseLockBackendTestMixin:
     def setUp(self):
-        super(FileLockTestCase, self).setUp()
+        super().setUp()
         self.locking_backend = import_string(dotted_path=self.backend_string)
 
     def test_exclusive(self):
@@ -71,8 +70,60 @@ class FileLockTestCase(BaseTestCase):
         # Cleanup
         lock_2.release()
 
+    def test_long_names(self):
+        # This test always passed on SQLite
+        # https://www.sqlite.org/datatype3.html, 3.1.1. Affinity Name Examples
+        # The `name` field will be created as a text field with no size limit.
+        self.locking_backend.acquire_lock(name=TEST_LOCK_LONG_NAME)
 
-class ModelLockTestCase(FileLockTestCase):
+    def test_purge(self):
+        self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=30)
+
+        # lock_1 not release and not expired, should raise LockError
+        with self.assertRaises(expected_exception=LockError):
+            self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        self.locking_backend.purge_locks()
+
+        # lock_1 not release but has expired, should not raise LockError
+        lock_2 = self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        # Cleanup
+        lock_2.release()
+
+
+class DefaultTimeoutTestMixin:
+    def setUp(self):
+        super().setUp()
+        self.locking_backend = import_string(dotted_path=self.backend_string)
+        setting_default_lock_timeout.set(value=1)
+        self.assertEqual(setting_default_lock_timeout.value, 1)
+
+    def test_default_timeout_expired(self):
+        self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=None)
+
+        # lock_1 not release and not expired, should raise LockError
+        with self.assertRaises(expected_exception=LockError):
+            self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        time.sleep(1.01)
+
+        # lock_1 not release but has expired, should not raise LockError
+        lock_2 = self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        # Cleanup
+        lock_2.release()
+
+
+class FileLockBackendTestCase(
+    BaseLockBackendTestMixin, DefaultTimeoutTestMixin, BaseTestCase
+):
+    backend_string = 'mayan.apps.lock_manager.backends.file_lock.FileLock'
+
+
+class ModelLockBackendTestCase(
+    BaseLockBackendTestMixin, DefaultTimeoutTestMixin, BaseTestCase
+):
     backend_string = 'mayan.apps.lock_manager.backends.model_lock.ModelLock'
 
 
@@ -80,5 +131,7 @@ class ModelLockTestCase(FileLockTestCase):
 @override_settings(
     LOCK_MANAGER_BACKEND_ARGUMENTS={'redis_url': 'redis://127.0.0.1:6379/0'}
 )
-class RedisLockTestCase(FileLockTestCase):
+class RedisLockBackendTestCase(
+    BaseLockBackendTestMixin, DefaultTimeoutTestMixin, BaseTestCase
+):
     backend_string = 'mayan.apps.lock_manager.backends.redis_lock.RedisLock'